server/requirements.txt
amqp==2.6.1
arabic-reshaper==2.1.3
arrow==0.15.5
asgiref==3.4.1
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
bitarray==1.2.2
boto3==1.16.12
botocore==1.19.63
cachetools==4.2.4
celery==4.4.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
cytoolz==0.11.0
decorator==5.0.9
Django==3.0.6
django-celery==3.1.17
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-cors-headers==3.4.0
django-extensions==3.1.3
django-health-check==3.16.5
django-jsonfield==1.4.1
django-model-utils==4.0.0
django-render-block==0.6
django-rest-swagger==2.2.0
django-s3-storage==0.13.4
django-silk==4.2.0
django-timezone-field==4.1.2
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
dnspython==2.1.0
email-validator==1.1.1
eth-abi==2.1.1
eth-account==0.5.4
eth-hash==0.3.1
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1
eth-typing==2.2.2
eth-utils==1.10.0
Faker==8.8.0
future==0.18.2
gevent==21.8.0
google-api-core==2.0.1
google-api-python-client==2.23.0
google-auth==2.2.1
google-auth-httplib2==0.1.0
google-oauth2-tool==0.0.3
googleapis-common-protos==1.53.0
gprof2dot==2021.2.21
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
hexbytes==0.2.1
html5lib==1.1
httplib2==0.19.1
idna==2.10
importlib-metadata==4.5.0
ipython==7.25.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
kombu==4.6.11
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
oauth2client==4.1.3
openapi-codec==1.3.2
parsimonious==0.8.1
parso==0.8.2
pdfkit==1.0.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
prompt-toolkit==3.0.19
protobuf==3.18.0
psycopg2-binary==2.8.5
ptyprocess==0.7.0
pyaml==20.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycryptodome==3.10.1
Pygments==2.9.0
PyJWT==1.7.1
pynt==0.8.2
pyparsing==2.4.7
PyPDF2==1.26.0
pysha3==1.0.2
python-bidi==0.4.2
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.9.1
python-googlegeocoder==0.4.0
python-http-client==3.3.2
pytz==2020.1
PyYAML==5.4.1
redis==3.5.3
reportlab==3.6.1
requests==2.25.1
rlp==2.0.1
rsa==4.7.2
s3transfer==0.3.7
sendgrid==6.4.1
sentry-sdk==1.3.1
silasdk==0.2.23
simplejson==3.17.2
six==1.16.0
soupsieve==2.2.1
sqlparse==0.3.1
starkbank-ecdsa==1.1.1
text-unidecode==1.3
toml==0.10.2
toolz==0.11.1
traitlets==5.0.5
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.5
uvicorn==0.14.0
vine==1.3.0
wcwidth==0.2.5
webencodings==0.5.1
xhtml2pdf==0.2.5
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
